feat : new module: main_svc
AbsenceException.java
AdminRequestState.java
application.properties
application.properties
Category.java
CategoryAdminController.java
CategoryDeleteException.java
CategoryDto.java
CategoryMapper.java
CategoryPublicController.java
CategoryRepository.java
CategoryService.java
CategoryServiceImpl.java
Client.java
Compilation.java
CompilationAdminController.java
CompilationCriteria.java
CompilationDto.java
CompilationMapper.java
CompilationPublicController.java
CompilationRepository.java
CompilationService.java
CompilationServiceImpl.java
CompilationSpecification.java
docker-compose.yml
EmailExistsEmail.java
Endpoint.java
EndpointDto.java
Event.java
EventAdminController.java
EventCriteria.java
EventFullDto.java
EventMapper.java
EventPrivateController.java
EventPublicController.java
EventRepository.java
EventRequestStatusUpdateRequest.java
EventRequestStatusUpdateResult.java
EventService.java
EventServiceImpl.java
EventShortDto.java
EventSpecification.java
EventStartBefore.java
EventStartTimeValidator.java
EventStartValidator.java
EventState.java
EventWrongTimeException.java
EwmService.java
ExceptionsHandler.java
ExceptionsHandler.java
Location.java
NameExistsException.java
NewCategoryDto.java
NewCompilationDto.java
NewEventDto.java
NewUserRequest.java
ParticipationRequest.java
ParticipationRequestDto.java
ParticipationRequestFailException.java
pom.xml
pom.xml
pom.xml
RequestMapper.java
RequestPrivateController.java
RequestRepository.java
RequestService.java
RequestServiceImpl.java
RequestStatus.java
schema.sql
SortParameter.java
UpdateEventDto.java
User.java
UserAdminController.java
UserCriteria.java
UserDto.java
UserMapper.java
UserRepository.java
UserRequestState.java
UserService.java
UserServiceImpl.java
UserShortDto.java
UserSpecification.java
ValidationGroups.java
ViewStatsDto.java
WrongEventStateException.java

+++ ewm/src/main/java/ru/practicum/category/service/CategoryServiceImpl.java
@@ -35,6 +35,9 @@ public class CategoryServiceImpl implements CategoryService {
     @Override
     public void deleteCategory(Long categoryId) {
         categoryRepository.findById(categoryId).orElseThrow(() -> new AbsenceException("Category not exists"));
+        if (eventRepository.existsByCategoryId(categoryId)) {
+            throw new CategoryDeleteException("Category used in Events");
+        }
         categoryRepository.deleteById(categoryId);
     }
